requirements.txt
asgiref==3.8.1
chardet==5.2.0
crispy-bootstrap5==2024.10
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-forms-bootstrap==3.1.0
django-import-export==4.3.7
django-simple-history==3.8.0
django-tables2==2.7.5
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
dj-database-url==2.3.0
django-unfold==0.50.0
et-xmlfile==2.0.0
filelock==3.16.1
gunicorn==23.0.0
MarkupPy==1.14
numpy==1.23.5
odfpy==1.4.1
openpyxl==3.1.5
pillow==10.4.0
platformdirs==4.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
PyYAML==6.0.1
setuptools==75.6.0
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.8.2
xlrd==2.0.1
xlwt==1.3.0
dj-config-url~=0.1.1
python-dotenv~=1.0.1
requests~=2.32.3
statsmodels=0.14.4
pandas~=2.2.3
xhtml2pdf~=0.2.16
xlsxwriter==3.2.2